added test scenarios
