+++ src/db-tests/fixtures/availableAreaCalculationScenarios.csv
@@ -8,3 +8,6 @@ UPL1 - Multiple existing compatible and incompatible actions,SD6743,7268,UPL1,"[
 OFM3 - Multiple existing incompatible actions that are compatible between them,SD6743,7268,OFM3,"[{ ""actionCode"": ""SPM4"", ""areaSqm"": 122.9 }, { ""actionCode"": ""SAM1"", ""areaSqm"": 522.9 }]",0.54039576
 CSAM1 - One existing incompatible action with no land cover in common,SD6853,6707,CSAM1,"[{ ""actionCode"": ""WBD1"", ""areaSqm"": 175.07 }]",2.41323074
 CMOR1 - One existing incompatible action with land cover in common,SD7164,8058,CMOR1,"[{ ""actionCode"": ""WS1"", ""areaSqm"": 25328.2335  }]",3.12543301
+UPL1 - Multiple existing incompatible actions-Existing and new actions are on different land covers,SD7948,4156,UPL1,"[ {  ""actionCode"": ""OP1"" ,  ""areaSqm"": 2000 }, {  ""actionCode"": ""OP2"" ,  ""areaSqm"": 2000 }, {  ""actionCode"": ""OP3"" ,  ""areaSqm"": 2000 }]",3.28437473
+CMOR1 - Multiple existing incompatible actions with single stack on mixed land covers,SD8447,1509,CMOR1,"[ {  ""actionCode"": ""CHRW1"" ,  ""areaSqm"": 100000 }, {  ""actionCode"": ""CHRW2"" ,  ""areaSqm"": 80000 }, {  ""actionCode"": ""CHRW3"" ,  ""areaSqm"": 70000 }]",1.150572
+UPL1 - Multiple existing Incompatible actions with multiple stacks on mixed landcovers,SD4957,8307,UPL1,"[ {  ""actionCode"": ""CHRW1"" ,  ""areaSqm"": 60000 }, {  ""actionCode"": ""CHRW2"" ,  ""areaSqm"": 40000 }, {  ""actionCode"": ""CHRW3"" ,  ""areaSqm"": 20000 }, {  ""actionCode"": ""PRF1"" ,  ""areaSqm"": 20000 }, {  ""actionCode"": ""PRF2"" ,  ""areaSqm"": 40000 }, {  ""actionCode"": ""GRH6"" ,  ""areaSqm"": 60000 }, {  ""actionCode"": ""GRH7"" ,  ""areaSqm"": 40000 }]",6.918519
